backend/requirements.txt
pyjwt
flask
flask-mail
pymongo[srv]
Flask-PyMongo
datetime
pathlib
bcrypt
requests
cryptography
virtualenv
pytest
google-api-python-client
pandas
numpy
scikit-learn
rake-nltk
validate_email
flask-cors
